django-site/todolist/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
dj-database-url==2.2.0
Django>=3.2,<5.0
djangorestframework==3.15.2
idna==3.8
psycopg2-binary==2.9.9
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.2
